Move towards quality:: and fast:: modules
