+++ rapidhash-bench/benches/rng.rs
@@ -28,7 +28,7 @@ pub fn bench_rapidhash(count: usize) -> Box<dyn FnMut(&mut Bencher)> {
             rand::random::<u64>()
         }, |i: u64| {
             let mut out = 0;
-            let mut rng = rapidhash::v3::RapidRng::seed_from_u64(i);
+            let mut rng = rapidhash::inner::RapidRng::seed_from_u64(i);
             for _ in 0..count {
                 out ^= rng.next_u64();
             }
@@ -44,7 +44,7 @@ pub fn bench_rapidhash_fast(count: usize) -> Box<dyn FnMut(&mut Bencher)> {
         }, |mut i: u64| {
             let mut out = 0;
             for _ in 0..count {
-                out ^= rapidhash::v3::rapidrng_fast(&mut i);
+                out ^= rapidhash::inner::rapidrng_fast(&mut i);
             }
             out
         }, criterion::BatchSize::SmallInput);
@@ -58,7 +58,7 @@ pub fn bench_rapidhash_time(count: usize) -> Box<dyn FnMut(&mut Bencher)> {
         }, |mut i: u64| {
             let mut out: u64 = 0;
             for _ in 0..=count {
-                out ^= rapidhash::v3::rapidrng_time(&mut i);
+                out ^= rapidhash::inner::rapidrng_time(&mut i);
             }
             out
         }, criterion::BatchSize::SmallInput);
