+++ rapidhash/src/fast.rs
@@ -1,5 +1,6 @@
 //! RapidHasher with a focus on hashing and data structure speed.
 //!
+//! This is a specific instantiation of the [rapidhash::inner] module with the following settings:
 //! - `AVALANCHE` is disabled.
 //! - `FNV` is enabled.
 //! - `COMPACT` is disabled.
@@ -10,10 +11,10 @@ const FNV: bool = true;
 const COMPACT: bool = false;
 const PROTECTED: bool = false;
 
-use crate::v3;
+use crate::inner;
 
 // TODO: random state etc.
-pub type RapidHasher = v3::RapidHasher<AVALANCHE, FNV, COMPACT, PROTECTED>;
-pub type RapidBuildHasher = v3::RapidBuildHasher<AVALANCHE, FNV, COMPACT, PROTECTED>;
-pub type RapidHashMap<K, V> = v3::RapidHashMap<K, V, AVALANCHE, FNV, COMPACT, PROTECTED>;
-pub type RapidHashSet<K> = v3::RapidHashSet<K, AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RapidHasher = inner::RapidHasher<AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RapidBuildHasher = inner::RapidBuildHasher<AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RapidHashMap<K, V> = inner::RapidHashMap<K, V, AVALANCHE, FNV, COMPACT, PROTECTED>;
+pub type RapidHashSet<K> = inner::RapidHashSet<K, AVALANCHE, FNV, COMPACT, PROTECTED>;
