+++ rapidhash/src/inner/rapid_const.rs
