+++ rapidhash/src/lib.rs
@@ -11,25 +11,14 @@
 mod mix;
 mod read;
 
-/// Rapidhash V1 algorithm implementations. Only exposed with the `versioned` feature.
-#[cfg(any(feature = "v1", docsrs))]
+/// Rapidhash V1 algorithm implementations.
 pub mod v1;
-/// Rapidhash V2 algorithm implementations (default). Only exposed with the `versioned` feature.
-#[cfg(any(feature = "v2", docsrs))]
+/// Rapidhash V2 algorithm implementations.
 pub mod v2;
-/// Rapidhash V3 algorithm implementations (default). Only exposed with the `versioned` feature.
-#[cfg(any(feature = "v3", docsrs))]
-pub mod v3;
 
-/// Privately load v2 when exporting default aliases.
-#[cfg(all(feature = "vlatest", not(any(feature = "v2", docsrs))))]
-pub(crate) mod v2;
-
-// #[cfg(any(feature = "vlatest", docsrs))]
-// mod latest;
-//
-// #[cfg(any(feature = "vlatest", docsrs))]
-// pub use latest::*;
+/// Rapidhash rust trait implementations.
+pub mod inner;
 
 pub mod fast;
 pub mod quality;
+pub mod rng;
